+++ packages/js-client-grpc/src/client-version.ts
@@ -1,4 +1,4 @@
-export const PACKAGE_VERSION = '1.12.0';
+export const PACKAGE_VERSION = '1.13.0';
 
 interface Version {
     major: number;