+++ packages/js-client-rest/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @qdrant/js-client-rest
 
+## 1.13.0
+
+### Minor Changes
+
+-   Qdrant v1.13.0 API
+
 ## 1.12.0
 
 ### Minor Changes